PositionApi => Position
Continues the series of Scaladoc-driven optimizations to scala-reflect.jar.
